pyproject.toml
[tool.poetry]
name = "snippet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.112.2"}
pydantic-settings = "^2.4.0"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
python-jose = {extras = ["cryptogrphy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
